resume_parser/requirements.txt
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
Django==2.2.10
django-crispy-forms==1.7.2
docx2txt==0.9
gunicorn==23.0.0
huggingface-hub
idna==3.11
llama-cpp-python
pdfminer.six==20251107
pycparser==2.23
pycryptodome==3.23.0
pyrsistent==0.20.0
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
urllib3==2.6.2
